Changed location of is_windows and removed every reference of is_windows from python_ebpf
